Update Game persistence between refreshes
Update the Tic Tac Toe Game to persist the board state between board
refreshes, including redrawing the final move. This required the board's
state was being track within the game state and not just the HTML board,
which required tweaking of values within the application to utilize this
new board state, as well as a new method to draw the previous board
state on refresh.